# api/openapi.yaml
openapi: 3.0.3
info:
  title: iRODS REST API
  version: 0.1.0
  description: |
    REST-ful iRODS running on GoLang.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/user:
    get:
      operationId: getUsers
      tags: [user]
      summary: List iRODS users
      description: |
        Lists visible iRODS user principals in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter for autocomplete.
      responses:
        "200":
          description: Users visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCollectionResponse"
        "400":
          description: Invalid user list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUser
      tags: [user]
      summary: Create one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make creation repeatable
        when the existing principal is already the requested user type. The operation
        remains a conflict when the existing principal is a group or when the existing
        user type differs from the requested type.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing user with the requested type
            as success instead of returning 409. A type mismatch or non-user
            principal with the same name remains a conflict. If a password is supplied
            for an existing matching user, the password is applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Desired user state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/user/{user_name}:
    get:
      operationId: getUser
      tags: [user]
      summary: Fetch one iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while reading the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putUser
      tags: [user]
      summary: Update one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin`.
        Synchronization clients may use `reconcile=true` as a constrained upsert:
        a missing user is created only when the request includes `type`. Password-only
        updates still require the user to exist, and existing non-user principals or
        incompatible types remain guarded by normal error handling.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Create the user when missing if the request
            includes a target `type`; password-only updates still return 404 when
            the user is absent. Reconciliation does not bypass rodsadmin authorization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid update request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUser
      tags: [user]
      summary: Delete one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` so deleting an already absent
        user is reported as success. The operation does not delete groups or other
        non-user principals with the same name.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat a missing user, or an existing non-user
            principal with this name, as already absent and return success.
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    ActionLink:
      type: object
      required: [href]
      properties:
        href:
          type: string
        method:
          type: string
...
    User:
      type: object
      required: [id, name, zone, type]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        links:
          $ref: "#/components/schemas/UserLinks"
    UserLinks:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/ActionLink"
        update:
          $ref: "#/components/schemas/ActionLink"
        delete:
          $ref: "#/components/schemas/ActionLink"
        avus:
          $ref: "#/components/schemas/ActionLink"
        create_avu:
          $ref: "#/components/schemas/ActionLink"
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    UserCollectionResponse:
      type: object
      required: [users, count, zone, links]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        count:
          type: integer
        zone:
          type: string
        type:
          type: string
          enum: [rodsuser, rodsadmin, groupadmin]
        prefix:
          type: string
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
            create:
              $ref: "#/components/schemas/ActionLink"
...
    UserCreateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          description: iRODS username to create.
        type:
          type: string
          enum: [rodsuser, rodsadmin]
          description: iRODS user type for the new user.
        password:
          type: string
          format: password
          description: Optional initial iRODS user password.
    UserUpdateRequest:
      type: object
      description: At least one of `type` or `password` is required.
      properties:
        type:
          type: string
          enum: [rodsuser, rodsadmin]
          description: New iRODS user type.
        password:
          type: string
          format: password
          description: New iRODS user password.
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          description: Optional field-level validation messages keyed by request field name.
          additionalProperties:
            type: string